feat: agregar test exhaustivo de endpoints con Jest y supertest

+++ src/core/application/usecase/product/DeleteProductUseCase.ts
@@ -8,20 +8,25 @@ import { BaseDeleteUseCase } from '../../base/BaseUseCase';
 import { IProductRepository } from '../../../domain/repository/ProductRepository';
 import { LoggerWrapperInterface } from '../../interface/LoggerWrapperInterface';
 import { Product } from '../../../domain/entity/Product';
+import { ServiceResult } from '../../../../infrastructure/services/base/ServiceTypes';
 
-export class DeleteProductUseCase extends BaseDeleteUseCase {
+export class DeleteProductUseCase extends BaseDeleteUseCase<Product> {
   constructor(
     private productRepository: IProductRepository,
     logger: LoggerWrapperInterface
   ) {
-    super(logger, { action: 'DELETE_PRODUCT', entityName: 'Producto' });
+    super(logger, { action: 'DELETE_PRODUCT', entityName: 'Product' });
   }
 
-  protected async findById(id: number): Promise<Product | null> {
+  protected async findEntityById(id: number): Promise<ServiceResult<Product>> {
     return this.productRepository.findById(id);
   }
 
-  protected async deleteEntity(id: number): Promise<void> {
-    await this.productRepository.delete(id);
+  protected async performDelete(id: number): Promise<ServiceResult<void>> {
+    const product = await this.findEntityById(id);
+    if (product.success && product.data && product.data.quantity > 0) {
+      throw new Error('Cannot delete a product with available stock');
+    }
+    return this.productRepository.delete(id);
   }
 }
